- Reduce preciptitation map to 32x32, 16x16 and 8x8 (use average values) - Normalize values

+++ server/InnAiServer/InnAiServer/Services/IRainRadarService.cs
@@ -6,7 +6,7 @@ public interface IRainRadarService
 {
     public Task<RainRadar[]> GetLastAsync(int count);
     public Task<byte[]> GetRadarImageAsync(string radarId);
-    public Task<int[,]> GetRadarImageValuesAsync(string radarId);
+    public Task<double[,]> GetRadarImageValuesAsync(string radarId);
     // public Task DownloadAndStoreLatestRadarImagesAsync();
 
     public Task LoadMonthAsync(int year, int month);
